edit_file
this script wrote for one of requirements, where it needs to find latest txt file [i.e., file generated today as fille was only generating once per day]. and remove 2nd and 2nd last line as those contains unnecessary pipes and remove extra pipes from first and last line and then save file to another location so it could publish to client. 
we need to remove those as with those extra pipe client's system is not consuming the file data and our inhouse file generator do not have option to do this.