search_file_n_move
this is the code for search_file_n_move

this is to search file in one location if file is available then move it to another location from where it will publish to client. 
this is done to reduce manual publishing of nearly 200 files each month.
the code will pick up file name from details file and search it in location where all publishing files are downloaded then make a list of available files then check if publishing path is available or not then move the file from source to destination.
all 200 files needs to move into 200 locations once it moved it will rename the file name with new name with date in it. 
once all done it will provide summary with how many files moved , how many files not found , how many destination paths are not available